Update GZ3D tutorial notebook
Karen has gone through and cleaned up some of the explanations and added a few links in places.

As part of this update we also fixed a bug in the `polar_plot` code where labels were not being used in the scatter plot calls.  The examples in the notebook now explicitly create figure legends to ensure this is working as expected.